Explicitly depend on libz4-dev with framing support (lz4frame.h)
As of librdkafka 0.9.5, libz4 is no longer an optional dependency.

The build system first examines the system's liblz4 for framing support
and if that fails, it fallbacks to the embedded version.

Since Debian cannot use the embedded library we explicitly depend
on a recent enough liblz4-dev.

Note that this affects builds for Debian Jessie (oldstable). For those
interested in building local packages for jessie, the suggested way is
to drop the B-Depends, the build system will then use the embedded
version. At least until liblz4-dev is backported to jessie-backports.